requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
streamlit
pandas
plotly
